backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
beautifulsoup4
requests
pydantic
langchain-core
langchain-community
langchain-google-genai
python-dotenv
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart

